academlo-django-002/src/requirements.txt
Django==2.2.24
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
